Fixed bug

+++ WebScraper/ScrapyProject/ScrapyProject/dbqueue.py
@@ -27,19 +27,22 @@ class DBQueue:
         self.cursor.execute("SELECT count(1) FROM queue WHERE url=%s", (item.url, ))
         exists = self.cursor.fetchone()[0]
 
-        if exists == '0':
+        if exists == 0:
             self.cursor.execute("INSERT INTO queue (priority, url, meta) VALUES (%s, %s, %s)",
-                            (priority, item.url, json.dumps(item.meta)))
+                                (priority, item.url, json.dumps(item.meta)))
             self.size += 1
 
     def pop(self):
         self.cursor.execute("SELECT id, priority, url, meta FROM queue ORDER BY priority desc, id asc LIMIT 1")
         result = self.cursor.fetchone()
-        # print(repr(result))
-        print("Deleting id %d" % result[0])
-        self.cursor.execute("DELETE FROM queue WHERE id=%s", (result[0],))
-        self.size -= 1
-        return scrapy.Request(url=result[2], priority=int(result[1]), meta=json.loads(result[3]))
+        if result is not None:
+            # print(repr(result))
+            # print("Deleting id %d" % result[0])
+            self.cursor.execute("DELETE FROM queue WHERE id=%s", (result[0],))
+            self.size -= 1
+            return scrapy.Request(url=result[2], priority=int(result[1]), meta=json.loads(result[3]))
+        else:
+            return None
 
     def close(self):
         self.cursor.close()
